feat(website): use readFile architecture
