+++ .idea/frontmatter.iml
@@ -6,6 +6,15 @@
       <excludeFolder url="file://$MODULE_DIR$/pnpm-lock.yaml" />
       <excludeFolder url="file://$MODULE_DIR$/.vercel" />
       <excludeFolder url="file://$MODULE_DIR$/workspace/sync-trustwallet-assets/repo" />
+      <excludeFolder url="file://$MODULE_DIR$/packages/eip155-1-erc20/frontmatter" />
+      <excludeFolder url="file://$MODULE_DIR$/packages/eip155-10-erc20/frontmatter" />
+      <excludeFolder url="file://$MODULE_DIR$/packages/eip155-56-erc20/frontmatter" />
+      <excludeFolder url="file://$MODULE_DIR$/packages/eip155-137-erc20/frontmatter" />
+      <excludeFolder url="file://$MODULE_DIR$/packages/eip155-8453-erc20/frontmatter" />
+      <excludeFolder url="file://$MODULE_DIR$/packages/eip155-42161-erc20/frontmatter" />
+      <excludeFolder url="file://$MODULE_DIR$/packages/eip155-42220-erc20/frontmatter" />
+      <excludeFolder url="file://$MODULE_DIR$/packages/eip155-43114-erc20/frontmatter" />
+      <excludeFolder url="file://$MODULE_DIR$/packages/eip155-1313161554-erc20/frontmatter" />
       <excludePattern pattern="*.turbo" />
       <excludePattern pattern="*dist" />
     </content>
