+++ website/components/contented/FrontmatterImage.tsx
@@ -1,5 +1,5 @@
 'use client';
-import { decodeCaip19, FrontmatterIndex, requireCryptoFrontmatter } from 'crypto-frontmatter';
+import { FrontmatterIndex, requireCryptoFrontmatter } from 'crypto-frontmatter';
 import Image from 'next/image';
 import type { ReactElement } from 'react';
 
@@ -9,10 +9,9 @@ export function FrontmatterImage(props: { frontmatter: FrontmatterIndex; type: '
     return <></>;
   }
 
-  const [caip2, type] = decodeCaip19(props.frontmatter.path);
   return (
     <Image
-      src={requireCryptoFrontmatter(caip2, type, image.path)}
+      src={requireCryptoFrontmatter(props.frontmatter.caip2, props.frontmatter.namespace, image.path)}
       alt={`${props.frontmatter.fields.symbol} ${props.type}`}
       width={image.size.width}
       height={image.size.height}
