+++ website/package.json
@@ -35,7 +35,7 @@
     "dayjs": "^1.11.10",
     "framer-motion": "10.16.4",
     "html-to-react": "^1.7.0",
-    "next": "14.0.1",
+    "next": "14.0.2",
     "react": "18.2.0",
     "react-dom": "18.2.0",
     "sharp": "^0.32.6",
