+++ workspace/eslint-config-next/package.json
@@ -3,7 +3,7 @@
   "private": true,
   "main": "./index.js",
   "dependencies": {
-    "@next/eslint-plugin-next": "14.0.1",
+    "@next/eslint-plugin-next": "14.0.2",
     "eslint-config-prettier": "^9.0.0",
     "eslint-plugin-import": "^2.29.0",
     "eslint-plugin-react": "^7.33.2",
